chore: Adjust code temporarily to enable build
