+++ EngineSIU/EngineSIU/Engine/Source/Editor/ViewerEditor/ViewerEditor.cpp
@@ -28,55 +28,55 @@ void ViewerEditor::CreateViewerWindow()
     }
 
 }
-
-void ViewerEditor::RenderSkeletalMesh(USkeletalMesh* SkeletalMeshComp) const
-{
-    ImGui::PushStyleColor(ImGuiCol_Header, ImVec4(0.1f, 0.1f, 0.1f, 1.0f));
-    if (ImGui::TreeNodeEx("Static Mesh", ImGuiTreeNodeFlags_Framed | ImGuiTreeNodeFlags_DefaultOpen)) // 트리 노드 생성
-    {
-        ImGui::Text("StaticMesh");
-        ImGui::SameLine();
-
-        FString PreviewName;
-        // TO-DO: USkeletalMeshComponent::GetSkeletalMesh()를 구현할 예정
-        // 다른 함수들도 많이 남았음.
-        if (SkeletalMeshComp->GetSkeletalMesh())
-        {
-            PreviewName = SkeletalMeshComp->GetSkeletalMesh()->GetRenderData()->DisplayName;
-        }
-        else
-        {
-            PreviewName = TEXT("None");
-        }
-
-        const TMap<FName, FAssetInfo> Assets = UAssetManager::Get().GetAssetRegistry();
-
-        if (ImGui::BeginCombo("##StaticMesh", GetData(PreviewName), ImGuiComboFlags_None))
-        {
-            if (ImGui::Selectable(TEXT("None"), false))
-            {
-                SkeletalMeshComp->SetSkeletalMesh(nullptr);
-            }
-
-            for (const auto& Asset : Assets)
-            {
-                if (Asset.Value.AssetType == EAssetType::SkeletalMesh)
-                {
-                    if (ImGui::Selectable(GetData(Asset.Value.AssetName.ToString()), false))
-                    {
-                        FString MeshName = Asset.Value.PackagePath.ToString() + "/" + Asset.Value.AssetName.ToString();
-                        USkeletalMesh* SkeletalMesh = FObjManager::GetSkeletalMesh(MeshName.ToWideString());
-                        if (SkeletalMesh)
-                        {
-                            SkeletalMeshComp->SetSkeletalMesh(SkeletalMesh);
-                        }
-                    }
-                }
-            }
-            ImGui::EndCombo();
-        }
-
-        ImGui::TreePop();
-    }
-    ImGui::PopStyleColor();
-}
+//
+//void ViewerEditor::RenderSkeletalMesh(USkeletalMesh* SkeletalMeshComp) const
+//{
+//    ImGui::PushStyleColor(ImGuiCol_Header, ImVec4(0.1f, 0.1f, 0.1f, 1.0f));
+//    if (ImGui::TreeNodeEx("Static Mesh", ImGuiTreeNodeFlags_Framed | ImGuiTreeNodeFlags_DefaultOpen)) // 트리 노드 생성
+//    {
+//        ImGui::Text("StaticMesh");
+//        ImGui::SameLine();
+//
+//        FString PreviewName;
+//        // TO-DO: USkeletalMeshComponent::GetSkeletalMesh()를 구현할 예정
+//        // 다른 함수들도 많이 남았음.
+//        if (SkeletalMeshComp->GetSkeletalMesh())
+//        {
+//            PreviewName = SkeletalMeshComp->GetSkeletalMesh()->GetRenderData()->DisplayName;
+//        }
+//        else
+//        {
+//            PreviewName = TEXT("None");
+//        }
+//
+//        const TMap<FName, FAssetInfo> Assets = UAssetManager::Get().GetAssetRegistry();
+//
+//        if (ImGui::BeginCombo("##StaticMesh", GetData(PreviewName), ImGuiComboFlags_None))
+//        {
+//            if (ImGui::Selectable(TEXT("None"), false))
+//            {
+//                SkeletalMeshComp->SetSkeletalMesh(nullptr);
+//            }
+//
+//            for (const auto& Asset : Assets)
+//            {
+//                if (Asset.Value.AssetType == EAssetType::SkeletalMesh)
+//                {
+//                    if (ImGui::Selectable(GetData(Asset.Value.AssetName.ToString()), false))
+//                    {
+//                        FString MeshName = Asset.Value.PackagePath.ToString() + "/" + Asset.Value.AssetName.ToString();
+//                        USkeletalMesh* SkeletalMesh = FObjManager::GetSkeletalMesh(MeshName.ToWideString());
+//                        if (SkeletalMesh)
+//                        {
+//                            SkeletalMeshComp->SetSkeletalMesh(SkeletalMesh);
+//                        }
+//                    }
+//                }
+//            }
+//            ImGui::EndCombo();
+//        }
+//
+//        ImGui::TreePop();
+//    }
+//    ImGui::PopStyleColor();
+//}
