--- EngineSIU/EngineSIU/SkinnedMeshComponent.cpp
@@ -1,1 +1,0 @@
-#include "SkinnedMeshComponent.h"
