+++ EngineSIU/EngineSIU/imgui.ini
@@ -4,8 +4,8 @@ Size=400,400
 Collapsed=0
 
 [Window][Outliner]
-Pos=1107,78
-Size=277,528
+Pos=1643,78
+Size=277,266
 Collapsed=0
 DockId=0x00000005,0
 
@@ -15,14 +15,14 @@ Size=270,624
 Collapsed=0
 
 [Window][Console]
-Pos=1107,608
-Size=277,353
+Pos=0,784
+Size=1920,353
 Collapsed=0
-DockId=0x00000006,1
+DockId=0x00000008,0
 
 [Window][Control Panel]
 Pos=0,0
-Size=1384,76
+Size=1920,76
 Collapsed=0
 DockId=0x00000001,0
 
@@ -32,10 +32,10 @@ Size=692,480
 Collapsed=0
 
 [Window][Engine Profiler]
-Pos=1107,78
-Size=277,528
+Pos=-8,600
+Size=1105,353
 Collapsed=0
-DockId=0x00000005,1
+DockId=0x00000003,1
 
 [Window][Menu]
 Pos=10,55
@@ -72,8 +72,8 @@ Size=147,116
 Collapsed=0
 
 [Window][Properties]
-Pos=1107,608
-Size=277,353
+Pos=1643,346
+Size=277,436
 Collapsed=0
 DockId=0x00000006,0
 
@@ -83,7 +83,7 @@ DockId=0x00000003
 
 [Window][WindowOverViewport_11111111]
 Pos=0,0
-Size=1384,961
+Size=1920,1137
 Collapsed=0
 
 [Window][SkeletalMesh Viewer]
@@ -93,11 +93,13 @@ Size=1535,1041
 Collapsed=0
 
 [Docking][Data]
-DockSpace       ID=0x08BD597D Window=0x1BBC0F80 Pos=8,31 Size=1384,961 Split=Y
-  DockNode      ID=0x00000001 Parent=0x08BD597D SizeRef=1384,76 Selected=0xF930105C
-  DockNode      ID=0x00000002 Parent=0x08BD597D SizeRef=1384,883 Split=X
-    DockNode    ID=0x00000003 Parent=0x00000002 SizeRef=1105,883 CentralNode=1
-    DockNode    ID=0x00000004 Parent=0x00000002 SizeRef=277,883 Split=Y
-      DockNode  ID=0x00000005 Parent=0x00000004 SizeRef=277,528 Selected=0xD712096B
-      DockNode  ID=0x00000006 Parent=0x00000004 SizeRef=277,353 Selected=0x8C72BEA8
+DockSpace         ID=0x08BD597D Window=0x1BBC0F80 Pos=0,23 Size=1920,1137 Split=Y
+  DockNode        ID=0x00000007 Parent=0x08BD597D SizeRef=1384,606 Split=Y
+    DockNode      ID=0x00000001 Parent=0x00000007 SizeRef=1384,76 Selected=0xF930105C
+    DockNode      ID=0x00000002 Parent=0x00000007 SizeRef=1384,883 Split=X
+      DockNode    ID=0x00000003 Parent=0x00000002 SizeRef=1105,883 CentralNode=1
+      DockNode    ID=0x00000004 Parent=0x00000002 SizeRef=277,883 Split=Y
+        DockNode  ID=0x00000005 Parent=0x00000004 SizeRef=277,334 Selected=0x66BFF854
+        DockNode  ID=0x00000006 Parent=0x00000004 SizeRef=277,547 Selected=0x8C72BEA8
+  DockNode        ID=0x00000008 Parent=0x08BD597D SizeRef=1384,353 Selected=0xEA83D666
